\name{BHMSMAfMRI-package}

\alias{BHMSMAfMRI}

\docType{package}

\title{Bayesian Hierarchical Multi-Subject Multiscale Analysis of Functional MRI Data}

\description{Performs BHMSMA (Sanyal & Ferreira, 2012) of fMRI data using wavelet based prior that borrows strength across subjects and returns posterior smoothed versions of the fMRI data and samples from the posterior distribution. Currently considers 2D slices only.
}

\details{
\tabular{ll}{
Package: \tab BHMSMAfMRI\cr
Type: \tab Package\cr
Version: \tab 1.3\cr
Date: \tab 2019-05-15\cr
License: \tab GPL (>= 2)\cr
}
%~~ An overview of how to use the package, including the most important ~~
%~~ functions ~~
}

\author{Nilotpal Sanyal <nsanyal@stanford.edu>, Marco Ferreira <marf@vt.edu>}

\references{
Sanyal, Nilotpal, and Ferreira, Marco A.R. (2012). Bayesian hierarchical multi-subject multiscale analysis of functional MRI data. Neuroimage, 63, 3, 1519-1531.
}

%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
%\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
